services/onboarding_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
asyncpg==0.29.0
pika==1.3.2
requests==2.31.0

